+++ album_ids.txt
@@ -1,4 +1,3 @@
-# album_ids:
 4aawyAB9vmqN3uQ7FjRGTy
 2ANVost0y2y52ema1E9xAZ
 1NAmidJlEaVgA3MpcPFYGq
@@ -10,5 +9,3 @@
 2ODvWsOgouMbaA5xf0RkJe
 1ATL5GLyefJaxhQzSPVrLX
 0S0KGZnfBGSIssfF54WSJh
-
-
