+++ artist_ids.txt
@@ -1,4 +1,3 @@
-# artist_ids:
 0TnOYISbd1XYRBk9myaseg
 1uNFoZAHBGtllmzznpCI3s
 4dpARuHxo51G3z768sgnrY
